+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/__tests__/checkoutConfiguration.test.js
@@ -1,9 +1,9 @@
 /**
  * @jest-environment jsdom
  */
-const store = require('../../../../../config/store');
+const store = require('../../../../../../config/store');
 const helpers = require('../helpers');
-const httpClient = require('../../../js/commons/httpClient');
+const httpClient = require('../../../../js/commons/httpClient');
 const {
   setCheckoutConfiguration,
 } = require('../checkoutConfiguration');
@@ -21,7 +21,7 @@ let applepay;
 let klarna;
 let querySelector;
 
-jest.mock('../../commons/httpClient', () => ({
+jest.mock('../../../commons/httpClient', () => ({
   httpClient: jest.fn().mockImplementation(() => ({
     balance: 100,
       resultCode: 'invalid'
