+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/__tests__/helpers.test.js
