+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/__tests__/renderGiftcardComponent.test.js
@@ -7,10 +7,10 @@ const {
   renderGiftCardSelectForm,
   showGiftCardWarningMessage,
 } = require('../giftcards');
-const store = require('../../../../../config/store');
+const store = require('../../../../../../config/store');
 
-jest.mock('../../commons');
-jest.mock('../../../../../config/store');
+jest.mock('../../../commons');
+jest.mock('../../../../../../config/store');
 
 beforeEach(() => {
   document.body.innerHTML = `
