+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/__tests__/renderPaymentMethod.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 const { renderPaymentMethod } = require('../renderPaymentMethod');
-const store = require('../../../../../config/store');
+const store = require('../../../../../../config/store');
 
 let mount;
 describe('Render Payment Method', () => {
