+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/helpers.js
@@ -1,6 +1,6 @@
-const store = require('../../../../config/store');
-const constants = require('../../../../config/constants');
-const { httpClient } = require('../commons/httpClient');
+const store = require('../../../../../config/store');
+const constants = require('../../../../../config/constants');
+const { httpClient } = require('../../commons/httpClient');
 
 function assignPaymentMethodValue() {
   const adyenPaymentMethod = document.querySelector('#adyenPaymentMethodName');
