+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/paymentMethodsConfiguration/giftcards/giftcardsConfig.js
@@ -10,7 +10,7 @@ const {
   GIFTCARD,
   SUCCESS,
   NOTENOUGHBALANCE,
-} = require('../../../../../../config/constants');
+} = require('../../../../../../../config/constants');
 const { initializeCheckout } = require('../../renderGenericComponent');
 
 class GiftCardConfig {
