+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/renderGenericComponent.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-unsafe-optional-chaining */
-const store = require('../../../../config/store');
+const store = require('../../../../../config/store');
 const { renderPaymentMethod } = require('./renderPaymentMethod');
 const helpers = require('./helpers');
-const constants = require('../../../../config/constants');
+const constants = require('../../../../../config/constants');
 
 function setCheckoutConfiguration(checkoutOptions) {
   const setField = (key, val) => val && { [key]: val };
