+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/review-page/checkoutReviewButtons.js
@@ -1,4 +1,4 @@
-const helpers = require('../../adyen_checkout/helpers');
+const helpers = require('../../adyen/checkout/helpers');
 const { httpClient } = require('../../commons/httpClient');
 
 /**
