+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/checkout.js
@@ -1,7 +1,7 @@
 /* eslint-disable prefer-regex-literals, global-require */
 const processInclude = require('base/util');
 const baseCheckout = require('base/checkout/checkout');
-const adyenCheckout = require('./adyen_checkout');
+const adyenCheckout = require('./adyen/checkout');
 
 $(document).ready(() => {
   processInclude(baseCheckout);
