+++ src/cartridges/app_adyen_SFRA/cartridge/templates/default/account/payment/paymentForm.isml
@@ -8,7 +8,7 @@
         <isscript>
             var assets = require('*/cartridge/scripts/assets.js');
             assets.addCss('/css/adyenCss.css');
-            assets.addJs('/js/adyen_account/adyenAccount.js');
+            assets.addJs('/js/adyen/account/adyenAccount.js');
         </isscript>
         <script type="text/javascript">
             window.Configuration = {
